TimedTransform takes less snapshots
Instead of just taking snapshots after an amount of seconds, now it monitors the velocity (position, rotation, and scale) and makes more informed decisions on when to take a snapshot.

Rotation currently requires snapshots to be taken after an amount of seconds like the old way. Might be possible to handle rotation entirely separately where the snapshots hold a "rotation velocity" instead of just the rotation values so that interpolation isn't entirely messed up as it currently is.